move register trees to page
