+++ pages/register-trees.tsx
@@ -1,0 +1,14 @@
+import React, { ReactElement } from 'react'
+import RegisterTrees from '../src/features/user/UserProfile/components/RegisterTrees'
+
+interface Props {
+
+}
+
+export default function Register({ }: Props): ReactElement {
+    return (
+        <div>
+            <RegisterTrees />
+        </div>
+    )
+}
